+++ src/widgets/create-exhibition/api/uploadImage.ts
