+++ src/widgets/create-exhibition/model/exhibitionFormHandler.ts
@@ -1,9 +1,9 @@
-import axios from 'axios';
 import { toast } from 'react-toastify';
+import { convertAddressToCoordinates } from '../api/convertAddressToCoordinates';
+import { createExhibition } from '../api/createExhibition';
+import { uploadImage } from '../api/uploadImage';
 import { ExhibitionFormData } from '../types/type';
 import { handleChageDate } from './changeDate';
-import { convertAddressToCoordinates } from './convertAddressToCoordinates';
-import { uploadImage } from './uploadImage';
 
 export const handleExhibitionFormSubmit = async (data: ExhibitionFormData) => {
   const coordinates = await convertAddressToCoordinates(data.address);
@@ -21,22 +21,14 @@ export const handleExhibitionFormSubmit = async (data: ExhibitionFormData) => {
     return;
   }
 
-  try {
-    const response = await axios.post('/api/expo', {
-      title: data.title,
-      description: data.introduction,
-      startedDay,
-      finishedDay,
-      location: data.location,
-      coverImage: img,
-      x: lng,
-      y: lat,
-    });
-    console.log('서버 응답:', response);
-
-    toast.success('Data successfully posted!');
-  } catch (error) {
-    console.error('Error:', error);
-    toast.error('Failed to post data');
-  }
+  await createExhibition({
+    title: data.title,
+    description: data.introduction,
+    startedDay,
+    finishedDay,
+    location: data.location,
+    coverImage: img,
+    x: lng,
+    y: lat,
+  });
 };
